Move Hooks and SetHooks functions to new file
